This patch fixes all warnings detected on x86_64.  Most fixes are trivial, such as casts from size_t and ptrdiff_t to int.
ath_hal_gettxbuf() returns a 32-bit address, so %p is not good to
display it - it's better to use explicit format 0x%x.

+++ trunk/net80211/ieee80211_output.c
@@ -1336,7 +1336,7 @@ ieee80211_setup_wpa_ie(struct ieee80211vap *vap, u_int8_t *ie)
 	ie[1] = frm - ie - 2;
 	KASSERT(ie[1]+2 <= sizeof(struct ieee80211_ie_wpa),
 		("WPA IE too big, %u > %u",
-		ie[1]+2, sizeof(struct ieee80211_ie_wpa)));
+		ie[1]+2, (int)sizeof(struct ieee80211_ie_wpa)));
 	return frm;
 #undef ADDSHORT
 #undef ADDSELECTOR
@@ -1419,7 +1419,7 @@ ieee80211_setup_rsn_ie(struct ieee80211vap *vap, u_int8_t *ie)
 	ie[1] = frm - ie - 2;
 	KASSERT(ie[1]+2 <= sizeof(struct ieee80211_ie_wpa),
 		("RSN IE too big, %u > %u",
-		ie[1]+2, sizeof(struct ieee80211_ie_wpa)));
+		ie[1]+2, (int)sizeof(struct ieee80211_ie_wpa)));
 	return frm;
 #undef ADDSELECTOR
 #undef ADDSHORT
